feat: move bank to c++ and add bank commands (#1371)
Move bank to C++ to avoid queries in lua. Adds a new `Bank` class that
can be used with either `Player` or `Guild` to manage transactions.

Also adds a new file `bank.lua` with a few simple commands to manage
your bank account remotely, this can be disabled in that same file.

Please delete options that are not relevant.

  - [x] New feature (non-breaking change which adds functionality)

I tried bank NPCs for all kinds of transactions; they weren't actually
working super well for guilds before but now they do. The hireling
BANKER skill was broken before this, I will fix that in a separate PR.